Add geometry utility

+++ src/player.cpp
@@ -11,8 +11,8 @@ namespace player {
 void setup() {
   player_info.w = 50;
   player_info.h = 100;
-  player_info.position = {SCREEN_WIDTH/2 - player_info.w/2,
-                          SCREEN_HEIGHT/2 - player_info.h/2};
+  player_info.position = {(float) SCREEN_WIDTH/2 - player_info.w/2,
+                          (float) SCREEN_HEIGHT/2 - player_info.h/2};
   player_info.direction = DOWN;
   player_info.speed = 100;
